project/server/requirements.txt
azure-ai-formrecognizer 
azure-core
flask
flask_cors




